pyproject.toml
[project]
name = "blogscraper"
version = "0.1.0"
description = "A tool for scraping blog posts from WordPress and Substack blogs and generating PDFs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "weasyprint>=60.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
blogscraper = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scrapers", "pdf_generator"]

[tool.uv]
dev-dependencies = []

